Major updates to lots of systems, including Destiny. Many missing and wrong consts added/corrected in packet_types.h some xmlp files updated to match client code.
fixed some corp problems in following files -
client, corp, account, config, marketproxy.

updated role intergers in many files.
lost few new updates for sql file....will find and update later.

+++ src/eve-server/character/CharUnboundMgrService.h
@@ -44,7 +44,7 @@ public:
     CharUnboundMgrService(PyServiceMgr* mgr);
     virtual ~CharUnboundMgrService();
 
-	void GetCharacterData(uint32 characterID, std::map<std::string, uint32> &characterDataMap);
+	void GetCharacterData(uint32 characterID, std::map<std::string, uint64> &characterDataMap);
 
 private:
     class Dispatcher;
